refactor: restructure api collector directories
Move modules in app and heartbeat package up 1 level. Also remove
Metrics class because it's currently not used.
