+++ fazbeat/api_collect/task/task_db_insert.py
@@ -6,16 +6,16 @@ from typing import TYPE_CHECKING
 
 from loguru import logger
 
-from fazbeat.api_collect.heartbeat.task._response_handler import ResponseHandler
-from fazbeat.api_collect.heartbeat.task.api_response_adapter import ApiResponseAdapter
+from fazbeat.api_collect.task._response_handler import ResponseHandler
+from fazbeat.api_collect.task.api_response_adapter import ApiResponseAdapter
 from fazutil.api.wynn.response.guild_response import GuildResponse
 from fazutil.api.wynn.response.online_players_response import OnlinePlayersResponse
 from fazutil.api.wynn.response.player_response import PlayerResponse
 from fazutil.heartbeat.task.itask import ITask
 
 if TYPE_CHECKING:
-    from fazbeat.api_collect.heartbeat.task.request_queue import RequestQueue
-    from fazbeat.api_collect.heartbeat.task.response_queue import ResponseQueue
+    from fazbeat.api_collect.task.request_queue import RequestQueue
+    from fazbeat.api_collect.task.response_queue import ResponseQueue
     from fazutil.api.wynn.wynn_api import WynnApi
     from fazutil.db.fazdb.fazdb_database import FazdbDatabase
